chore: move CustomerFormModal to features
